+++ lib/AL_USDMaya/AL/usdmaya/cmds/ProxyShapeCommands.h
@@ -264,6 +264,29 @@ private:
   MStatus redoIt() override;
 };
 
+
+//----------------------------------------------------------------------------------------------------------------------
+/// \brief  TranslatePrim
+/// \ingroup commands
+//----------------------------------------------------------------------------------------------------------------------
+class TranslatePrim
+  : public ProxyShapeCommandBase
+{
+  fileio::translators::TranslatorParameters tp;
+
+  nodes::ProxyShape* m_proxy;
+  SdfPathVector m_importPaths;
+  SdfPathVector m_teardownPaths;
+
+public:
+  AL_MAYA_DECLARE_COMMAND();
+private:
+  bool isUndoable() const override;
+  MStatus doIt(const MArgList& args) override;
+  MStatus redoIt() override;
+};
+
+
 /// \brief  This function will generate all of the MEL script menus, and the option boxes, for all of the proxy shape
 ///         commands.
 /// \ingroup commands
