+++ lib/AL_USDMaya/AL/usdmaya/nodes/ProxyShape.cpp
@@ -39,6 +39,7 @@
 #include "AL/usdmaya/nodes/TransformationMatrix.h"
 #include "AL/usdmaya/nodes/proxy/PrimFilter.h"
 #include "AL/usdmaya/Version.h"
+#include "AL/usd/utils/ForwardDeclares.h"
 
 #include "maya/MFileIO.h"
 #include "maya/MFnPluginData.h"
@@ -53,6 +54,8 @@
 #include "pxr/base/tf/fileUtils.h"
 #include "pxr/usd/ar/resolver.h"
 #include "pxr/usd/usd/stageCacheContext.h"
+#include "pxr/usdImaging/usdImaging/primAdapter.h"
+#include "pxr/usdImaging/usdImaging/meshAdapter.h"
 
 #include <algorithm>
 #include <iterator>
@@ -61,6 +64,15 @@ namespace AL {
 namespace usdmaya {
 namespace nodes {
 typedef void (*proxy_function_prototype)(void* userData, AL::usdmaya::nodes::ProxyShape* proxyInstance);
+
+MDagPath ProxyShape::parentTransform()
+{
+  MFnDagNode fn(thisMObject());
+  MDagPath proxyTransformPath;
+  fn.getPath(proxyTransformPath);
+  proxyTransformPath.pop();
+  return proxyTransformPath;
+}
 
 //----------------------------------------------------------------------------------------------------------------------
 void ProxyShape::serialiseTranslatorContext()
@@ -99,6 +111,7 @@ MObject ProxyShape::m_filePath = MObject::kNullObj;
 MObject ProxyShape::m_primPath = MObject::kNullObj;
 MObject ProxyShape::m_excludePrimPaths = MObject::kNullObj;
 MObject ProxyShape::m_populationMaskIncludePaths = MObject::kNullObj;
+MObject ProxyShape::m_excludedTranslatedGeometry = MObject::kNullObj;
 MObject ProxyShape::m_time = MObject::kNullObj;
 MObject ProxyShape::m_timeOffset = MObject::kNullObj;
 MObject ProxyShape::m_timeScalar = MObject::kNullObj;
@@ -176,6 +189,108 @@ UsdStagePopulationMask ProxyShape::constructStagePopulationMask(const MString &p
   return mask;
 }
 
+void ProxyShape::translatePrimPathsIntoMaya(
+    const SdfPathVector& importPaths,
+    const SdfPathVector& teardownPaths,
+    const fileio::translators::TranslatorParameters& param)
+{
+  TF_DEBUG(ALUSDMAYA_EVALUATION).Msg("ProxyShape:translatePrimPathsIntoMaya ImportSize='%d' TearDownSize='%d' \n",
+                                     importPaths.size(),
+                                     teardownPaths.size());
+
+  //Resolve SdfPathSet to UsdPrimVector
+  UsdPrimVector importPrims;
+  for(const SdfPath& path : importPaths)
+  {
+    UsdPrim prim = m_stage->GetPrimAtPath(path);
+    if(prim.IsValid())
+    {
+      importPrims.push_back(prim);
+    }
+    else
+    {
+      TF_DEBUG(ALUSDMAYA_EVALUATION).Msg("ProxyShape:translatePrimPathsIntoMaya Path for import '%s' resolves to an invalid prim\n", path.GetString().c_str());
+    }
+  }
+
+  translatePrimsIntoMaya(importPrims, teardownPaths, param);
+}
+
+void ProxyShape::translatePrimsIntoMaya(
+    const UsdPrimVector& importPrims,
+    const SdfPathVector& teardownPrims,
+    const fileio::translators::TranslatorParameters& param)
+{
+  TF_DEBUG(ALUSDMAYA_EVALUATION).Msg("ProxyShape:translatePrimsIntoMaya ImportSize='%d' TearDownSize='%d' \n", importPrims.size(), teardownPrims.size());
+
+  proxy::PrimFilter filter(teardownPrims, importPrims, this);
+  if(TfDebug::IsEnabled(ALUSDMAYA_TRANSLATORS))
+  {
+    std::cout << "new prims" << std::endl;
+    for(auto it : filter.newPrimSet())
+    {
+      std::cout << it.GetPath().GetText() << std::endl;
+    }
+    std::cout << "new transforms" << std::endl;
+    for(auto it : filter.transformsToCreate())
+    {
+      std::cout << it.GetPath().GetText() << std::endl;
+    }
+    std::cout << "updateable prims" << std::endl;
+    for(auto it : filter.updatablePrimSet())
+    {
+      std::cout << it.GetPath().GetText() << std::endl;
+    }
+    std::cout << "removed prims" << std::endl;
+    for(auto it : filter.removedPrimSet())
+    {
+      std::cout << it.GetText() << std::endl;
+    }
+  }
+
+  cmds::ProxyShapePostLoadProcess::MObjectToPrim objsToCreate;
+  if(!filter.transformsToCreate().empty())
+  {
+    cmds::ProxyShapePostLoadProcess::createTranformChainsForSchemaPrims(
+        this,
+        filter.transformsToCreate(),
+        parentTransform(),
+        objsToCreate);
+  }
+
+  context()->removeEntries(filter.removedPrimSet());
+
+  if(!filter.newPrimSet().empty())
+  {
+    cmds::ProxyShapePostLoadProcess::createSchemaPrims(this, filter.newPrimSet(), param);
+  }
+
+  if(!filter.updatablePrimSet().empty())
+  {
+    cmds::ProxyShapePostLoadProcess::updateSchemaPrims(this, filter.updatablePrimSet());
+  }
+
+  cleanupTransformRefs();
+
+  context()->updatePrimTypes();
+
+  // now perform any post-creation fix up
+  if(!filter.newPrimSet().empty())
+  {
+    cmds::ProxyShapePostLoadProcess::connectSchemaPrims(this, filter.newPrimSet());
+  }
+
+  if(!filter.updatablePrimSet().empty())
+  {
+    cmds::ProxyShapePostLoadProcess::connectSchemaPrims(this, filter.updatablePrimSet());
+  }
+
+  if(context()->isExcludedGeometryDirty())
+  {
+    TF_DEBUG(ALUSDMAYA_EVALUATION).Msg("ProxyShape:translatePrimsIntoMaya excluded geometry has been modified, reconstructing imaging engine \n");
+    constructGLImagingEngine();
+  }
+}
 //----------------------------------------------------------------------------------------------------------------------
 SdfPathVector ProxyShape::getPrimPathsFromCommaJoinedString(const MString &paths) const
 {
@@ -191,6 +306,7 @@ SdfPathVector ProxyShape::getPrimPathsFromCommaJoinedString(const MString &paths
       begin = iter + 1;
       iter = std::find(begin, end, ',');
     }
+
     result.push_back(SdfPath(std::string(begin, end)));
   }
   return result;
@@ -215,11 +331,15 @@ void ProxyShape::constructGLImagingEngine()
         delete m_engine;
       }
 
+      const SdfPathSet& translatedGeo = m_context->excludedGeometry();
       // combine the excluded paths
       SdfPathVector excludedGeometryPaths;
-      excludedGeometryPaths.reserve(m_excludedTaggedGeometry.size() + m_excludedGeometry.size());
+      excludedGeometryPaths.reserve(m_excludedTaggedGeometry.size() + m_excludedGeometry.size() + translatedGeo.size());
       excludedGeometryPaths.assign(m_excludedTaggedGeometry.begin(), m_excludedTaggedGeometry.end());
       excludedGeometryPaths.insert(excludedGeometryPaths.end(), m_excludedGeometry.begin(), m_excludedGeometry.end());
+      excludedGeometryPaths.insert(excludedGeometryPaths.end(),
+                                   translatedGeo.begin(),
+                                   translatedGeo.end());
 
       m_engine = new UsdImagingGLHdEngine(m_path, excludedGeometryPaths);
 
@@ -503,6 +623,7 @@ MStatus ProxyShape::initialise()
     m_primPath = addStringAttr("primPath", "pp", kCached | kReadable | kWritable | kStorable | kAffectsAppearance);
     m_excludePrimPaths = addStringAttr("excludePrimPaths", "epp", kCached | kReadable | kWritable | kStorable | kAffectsAppearance);
     m_populationMaskIncludePaths = addStringAttr("populationMaskIncludePaths", "pmi", kCached | kReadable | kWritable | kStorable | kAffectsAppearance);
+    m_excludedTranslatedGeometry = addStringAttr("excludedTranslatedGeometry", "etg", kCached | kReadable | kWritable | kStorable | kAffectsAppearance);
 
     m_complexity = addInt32Attr("complexity", "cplx", 0, kCached | kConnectable | kReadable | kWritable | kAffectsAppearance | kKeyable | kStorable);
     setMinMax(m_complexity, 0, 8, 0, 4);
@@ -628,63 +749,17 @@ void ProxyShape::onPrimResync(SdfPath primPath, SdfPathVector& previousPrims)
 
   AL_BEGIN_PROFILE_SECTION(ObjectChanged);
   MFnDagNode fn(thisMObject());
-  MDagPath dag_path;
-  fn.getPath(dag_path);
-  dag_path.pop();
+  MDagPath proxyTransformPath;
+  fn.getPath(proxyTransformPath);
+  proxyTransformPath.pop();
 
-  // find the new set of schema translator prims
-  std::vector<UsdPrim> newPrimSet = huntForNativeNodesUnderPrim(dag_path, primPath, translatorManufacture());
+  // find the new set of prims
+  UsdPrimVector newPrimSet = huntForNativeNodesUnderPrim(proxyTransformPath, primPath, translatorManufacture());
 
-  proxy::PrimFilter filter(previousPrims, newPrimSet, this);
+  // Remove prims that have disappeared and translate in new prims
+  translatePrimsIntoMaya(newPrimSet, previousPrims);
+
   previousPrims.clear();
-
-  if(TfDebug::IsEnabled(ALUSDMAYA_TRANSLATORS)){
-    std::cout << "new prims:" << std::endl;
-    for(auto it : filter.newPrimSet())
-    {
-      std::cout << it.GetPath().GetText() << std::endl;
-    }
-    std::cout << "new transforms:" << std::endl;
-    for(auto it : filter.transformsToCreate())
-    {
-      std::cout << it.GetPath().GetText() << std::endl;
-    }
-    std::cout << "updateable prims:" << std::endl;
-    for(auto it : filter.updatablePrimSet())
-    {
-      std::cout << it.GetPath().GetText() << std::endl;
-    }
-    std::cout << "removed prims:" << std::endl;
-    for(auto it : filter.removedPrimSet())
-    {
-      std::cout << it.GetText() << std::endl;
-    }
-  }
-
-  // create transforms first to increment ref counts and avoid deleting/recreating ones that will stay
-  cmds::ProxyShapePostLoadProcess::MObjectToPrim objsToCreate;
-  if(!filter.transformsToCreate().empty())
-    cmds::ProxyShapePostLoadProcess::createTranformChainsForSchemaPrims(this, filter.transformsToCreate(), dag_path, objsToCreate);
-
-  context()->removeEntries(filter.removedPrimSet());
-
-  if(!filter.newPrimSet().empty())
-    cmds::ProxyShapePostLoadProcess::createSchemaPrims(this, filter.newPrimSet());
-
-  if(!filter.updatablePrimSet().empty())
-    cmds::ProxyShapePostLoadProcess::updateSchemaPrims(this, filter.updatablePrimSet());
-
-  cleanupTransformRefs();
-
-  context()->updatePrimTypes();
-
-  // now perform any post-creation fix up
-  if(!filter.newPrimSet().empty())
-    cmds::ProxyShapePostLoadProcess::connectSchemaPrims(this, filter.newPrimSet());
-
-  if(!filter.updatablePrimSet().empty())
-    cmds::ProxyShapePostLoadProcess::connectSchemaPrims(this, filter.updatablePrimSet());
-
 
   TF_DEBUG(ALUSDMAYA_TRANSLATORS).Msg("ProxyShape::onPrimResync end:\n%s\n", context()->serialise().asChar());
 
@@ -971,7 +1046,8 @@ std::vector<UsdPrim> ProxyShape::huntForNativeNodesUnderPrim(
       continue;
     }
 
-    if(utils.isSchemaPrim(prim))
+    fileio::translators::TranslatorRefPtr trans = utils.isSchemaPrim(prim);
+    if(trans && trans->importableByDefault())
     {
       prims.push_back(prim);
     }
@@ -998,7 +1074,6 @@ void ProxyShape::variantSelectionListener(SdfNotice::LayersDidChange const& noti
   {
     return;
   }
-
 
   TF_FOR_ALL(itr, notice.GetChangeListMap())
   {
@@ -1199,7 +1274,6 @@ void ProxyShape::loadStage()
       // execute the post load process to import any custom prims
       cmds::ProxyShapePostLoadProcess::initialise(this);
       findTaggedPrims();
-
     AL_END_PROFILE_SECTION();
   }
