+++ plugin/AL_USDMayaTestPlugin/run_mayaplugin_tests.mel
@@ -1,3 +1,3 @@
 loadPlugin "AL_USDMayaTestPlugin";
-int $res = `AL_usdmaya_UnitTestHarness`;
+int $res = `AL_usdmaya_UnitTestHarness -f "TranslateCommand*"`;
 quit -f -exitCode `abs($res)`;
