+++ schemas/AL/usd/schemas/schema.usda.in
@@ -51,6 +51,7 @@ over "GLOBAL" (
         }
     }
 ) {}
+
 ###################################################################################################
 
 class "ALModelAPI"
@@ -88,3 +89,4 @@ class ALMayaReference "ALMayaReference"(
   )
 }
 
+###################################################################################################
