+++ translators/tests/CMakeLists.txt
@@ -8,5 +8,5 @@ enable_testing()
 add_test(
     NAME TestAdditionalTranslators
     WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
-    COMMAND testMayaRef.sh ${CMAKE_INSTALL_PREFIX} $<TARGET_FILE_DIR:plug>
+    COMMAND testTranslators.sh ${CMAKE_INSTALL_PREFIX} $<TARGET_FILE_DIR:plug>
     )
