+++ translators/tests/testMeshVariants.usda
@@ -1,0 +1,32 @@
+#usda 1.0
+
+
+def "TestVariantSwitch"(
+  add variantSets = ["MeshVariants"]
+)
+{
+    variantSet "MeshVariants" = {
+        "ShowMeshA" ()
+        {
+          def "MeshA"(
+            references = [@./sphere.usda@]
+          ){}
+        }
+        "ShowMeshB" ()
+        {
+          def "MeshB"(
+            references = [@./sphere.usda@]
+          ){}
+        }
+        "ShowMeshAnB" () 
+        {
+          def "MeshA"(
+            references = [@./sphere.usda@]
+          ){}
+          
+          def "MeshB"(
+            references = [@./sphere.usda@]
+          ){}
+        }
+    }
+}
