+++ translators/tests/testTranslators.sh
@@ -7,6 +7,4 @@ export LD_LIBRARY_PATH=$AL_USDMAYA_LOCATION/lib:$USD_LIBRARY_PATH:$LD_LIBRARY_PA
 export PYTHONPATH=$AL_USDMAYA_LOCATION/lib/python:$USD_LIBRARY_PATH/python:$PYTHONPATH
 export PATH=$MAYA_LOCATION/bin:$PATH
 
-echo "USD_LIBRARY_PATH ${USD_LIBRARY_PATH}"
-
-mayapy testMayaRef.py
+mayapy testTranslators.py
